% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draculR_counts.R
\name{draculR_parse_counts}
\alias{draculR_parse_counts}
\title{Parse miRNA counts to calculate draculR haemolysis score}
\usage{
draculR_parse_counts(
  counts_df,
  drop_miRs = c(""),
  verbose = FALSE,
  filterNum = 1
)
}
\arguments{
\item{counts_df}{Dataframe containing samples as columns and miRNAs as rows with raw counts}

\item{drop_miRs}{Vector of miRNAs to remove from the original 20 miRNA used in the calculation}

\item{verbose}{Boolean to show more/less information}

\item{filterNum}{Minimun number of counts for each sample}
}
\value{
A dataframe with the classification of each sample as Clear or Caution followin the original calculation by draculR
}
\description{
This is the minimun working code to calculate the haemolysis score for each sample.
}
