% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draculR_validate.R
\name{draculR_parse_file}
\alias{draculR_parse_file}
\title{Parse miRNA counts matrix file}
\usage{
draculR_parse_file(raw_data, sep_input = ",", verbose = FALSE)
}
\arguments{
\item{raw_data}{Absolute path to raw_Data counts}

\item{sep_input}{Separator: ",", "\t", ";", etc}

\item{verbose}{Boolean to show more/less information}
}
\value{
Dataframe containing samples as columns and miRNAs as rows with raw counts
}
\description{
This function parses the original raw counts data that requires to meet some criteria
of minimun count, colnames, etc
}
