% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draculR_validate.R
\name{validate_input}
\alias{validate_input}
\title{Wrapper function for validating miRNA counts matrix file}
\usage{
validate_input(
  raw_data,
  radio_separator = ",",
  verbose = FALSE,
  file_given = TRUE
)
}
\arguments{
\item{raw_data}{Absolute path to raw_Data counts}

\item{radio_separator}{Separator: ",", "\t", ";", etc}

\item{verbose}{Boolean to show more/less information}

\item{file_given}{Boolean to identify either raw_data is a file or a dataframe}
}
\value{
Boolean with either TRUE/FALSE if it meets the criteria
}
\description{
This function wraps other functions to check whether data meets the required criteria
of minimun count, colnames, etc
}
